# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Tolti AI Backend API
  description: |
    Sovereign, on-premise, multiplayer AI workbench backend.

    Every endpoint is scoped to a workspace and authorised by JWT + RBAC.
    Real-time collaboration runs over the WebSocket protocol documented separately.
  version: 0.1.0
  contact:
    name: Tolti AI
    url: https://tolti.local
servers:
  - url: http://localhost:3001
    description: Local dev backend
  - url: http://localhost/api
    description: Local dev via nginx proxy
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/auth/login:
    post:
      tags: [auth]
      security: []
      summary: Email/password login
      responses:
        '200': { description: Auth session }
        '401': { description: Invalid credentials }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT